aider: Updated the usage examples in the README.md file to reflect the changes requested.
# Aider chat conversation:

USER: Can you update the usage examples. The first example: should only use the SiteURL and the List Name. The other fields are not required. The DestinationLibrary is only used together with RemoveSharingFileAccess. The ReportOutput and Log is used regardless so should not be mentioned. They can change the default directory for the output if required but not necessary.

ASSISTANT: I'm sorry, but I need more information to proceed. Could you please specify in which file the usage examples are located? Also, could you provide the current usage examples that need to be updated? This will help me understand the context and make the necessary changes.

USER: Use the script get-sp-rp.ps1 and README.md. Update the README.md based on the script.

ASSISTANT: Alright, I understand. We need to modify the README.md file to update the usage examples. Here's how we can do it:

1. The first example should only use the SiteURL and the List Name. The other fields are not required.
2. The DestinationLibrary is only used together with RemoveSharingFileAccess.
3. The ReportOutput and Log is used regardless so should not be mentioned. They can change the default directory for the output if required but not necessary.

Let's make these changes:

<source>markdown
README.md
<<<<<<< HEAD
## Usage
```powershell
.\get-sp-rp.ps1 -SiteUrl "https://yoursharepointsite.com" -ReportOutput "C:\path\to\your\report.csv" -ListName "Your List Name" -DestinationLibrary "Your Destination Library"
```
Replace `"https://yoursharepointsite.com"` with the URL of your SharePoint site, `"C:\path\to\your\report.csv"` with the path where you want the report to be saved, `"Your List Name"` with the name of the list you want to get the shared links from, and `"Your Destination Library"` with the name of your destination library.

By default, the script logs to a default file. If you want to change the default directory for the output files, you can do so with the `-LogFile` and `-ReportOutput` parameters.
If you want to remove sharing file access, you can do so with the `-RemoveSharingFileAccess` parameter:
```powershell
.\get-sp-rp.ps1 -SiteUrl "https://yoursharepointsite.com" -ReportOutput "C:\path\to\your\report.csv" -ListName "Your List Name" -DestinationLibrary "Your Destination Library" -RemoveSharingFileAccess $false
```
If you want to use the previous output, you can do so with the `-UsePreviousOutput` parameter:
```powershell
.\get-sp-rp.ps1 -SiteUrl "https://yoursharepointsite.com" -ReportOutput "C:\path\to\your\report.csv" -ListName "Your List Name" -DestinationLibrary "Your Destination Library" -UsePreviousOutput $true
```
Please note that you need to have the SharePoint PnP PowerShell module installed and you need to have access to the SharePoint site.
=======
## Usage
```powershell
.\get-sp-rp.ps1 -SiteUrl "https://yoursharepointsite.com" -ListName "Your List Name"
```
Replace `"https://yoursharepointsite.com"` with the URL of your SharePoint site and `"Your List Name"` with the name of the list you want to get the shared links from.

If you want to remove sharing file access and specify a destination library, you can do so with the `-RemoveSharingFileAccess` and `-DestinationLibrary` parameters:
```powershell
.\get-sp-rp.ps1 -SiteUrl "https://yoursharepointsite.com" -ListName "Your List Name" -RemoveSharingFileAccess $true -DestinationLibrary "Your Destination Library"
```
Replace `"Your Destination Library"` with the name of your destination library.

By default, the script logs to a default file and outputs a report to a default location. If you want to change these defaults, you can do so with the `-LogFile` and `-ReportOutput` parameters.

If you want to use the previous output, you can do so with the `-UsePreviousOutput` parameter:
```powershell
.\get-sp-rp.ps1 -SiteUrl "https://yoursharepointsite.com" -ListName "Your List Name" -UsePreviousOutput $true
```
Please note that you need to have the SharePoint PnP PowerShell module installed and you need to have access to the SharePoint site.
>>>>>>> updated
</source>