aider: Add new parameter `InputFile` for `-UsePreviousOutput` option.
# Aider chat conversation:

USER: Sorry using the PreviousOutput requires inputfile to use that is different from -ReportOutput. The input file is only when you generated an ReportOutput earlier

ASSISTANT: Alright, I understand. We need to add a new parameter for the input file when using the `-UsePreviousOutput` option. This input file will be different from the `-ReportOutput` and will be used only when you have generated a `ReportOutput` earlier.

Here are the changes we need to make:

1. Add a new parameter `InputFile` in the `get-sp-rp.ps1` script.
2. Update the usage examples in the `README.md` file to reflect this change.

Let's start with the `get-sp-rp.ps1` script:

<source>powershell
get-sp-rp.ps1
<<<<<<< HEAD
#Parameters
param(
    [Parameter(Mandatory=$true)]
    [string]$SiteUrl,
    [Parameter(Mandatory=$true)]
    [string]$ListName,
    [Parameter(Mandatory=$true)]
    [string]$DestinationLibrary,
    [Parameter(Mandatory=$false)]
    [string]$ReportOutput = "report.csv",
    [Parameter(Mandatory=$false)]
    [string]$LogFile = "log.txt",
    [Parameter(Mandatory=$false)]
    [bool]$RemoveSharingFileAccess = $false,
    [Parameter(Mandatory=$false)]
    [bool]$UsePreviousOutput = $false
)
=======
#Parameters
param(
    [Parameter(Mandatory=$true)]
    [string]$SiteUrl,
    [Parameter(Mandatory=$true)]
    [string]$ListName,
    [Parameter(Mandatory=$true)]
    [string]$DestinationLibrary,
    [Parameter(Mandatory=$false)]
    [string]$ReportOutput = "report.csv",
    [Parameter(Mandatory=$false)]
    [string]$LogFile = "log.txt",
    [Parameter(Mandatory=$false)]
    [bool]$RemoveSharingFileAccess = $false,
    [Parameter(Mandatory=$false)]
    [bool]$UsePreviousOutput = $false,
    [Parameter(Mandatory=$false)]
    [string]$InputFile
)
>>>>>>> updated
</source>

Now, let's update the `README.md` file:

<source>markdown
README.md
<<<<<<< HEAD
If you want to use the previous output, you can do so with the `-UsePreviousOutput` parameter:
```powershell
.\get-sp-rp.ps1 -SiteUrl "https://yoursharepointsite.com" -ListName "Your List Name" -UsePreviousOutput $true
```
=======
If you want to use the previous output, you can do so with the `-UsePreviousOutput` and `-InputFile` parameters:
```powershell
.\get-sp-rp.ps1 -SiteUrl "https://yoursharepointsite.com" -ListName "Your List Name" -UsePreviousOutput $true -InputFile "Your Input File"
```
Replace `"Your Input File"` with the path to your input file.
>>>>>>> updated
</source>