aider: Updated README.md with more examples on the usage of the parameters and a note about using -DestinationLibrary with -RemoveSharingFileAccess.
# Aider chat conversation:

USER: how does get-sp-rp.ps1 use inputfile ?

ASSISTANT: The `get-sp-rp.ps1` script uses the `InputFile` parameter in the following way:

1. The `InputFile` parameter is optional. If it is provided, the script will read the file and process only those items. If it is not provided, the script will fetch all items.

2. The `InputFile` is read using the `Get-Content` cmdlet, which gets the content of the item at the location specified by the path, such as the text in a file. It reads the content one line at a time, returning each line as a separate string.

3. The `ForEach-Object` cmdlet is used to perform an operation on each item. In this case, it calls `Get-PnPListItem` for each line in the `InputFile`.

Here is the relevant code:

```powershell
if ($InputFile) {
    # If an input file is provided, read the file and process only those items
    $ListItems = Get-Content $InputFile | ForEach-Object {
        Get-PnPListItem -List $ListName -Id $_ -Fields "FileLeafRef", "FileRef", "File_x0020_Type"
    }
} else {
    # If no input file is provided, fetch all items
    $ListItems = Get-PnPListItem -List $ListName -PageSize 2000 -Fields "FileLeafRef", "FileRef", "File_x0020_Type"
}
```

So, if an `InputFile` is provided, the script will process only the items specified in that file. Each line in the file should represent an item ID. If no `InputFile` is provided, the script will process all items in the list.

USER: Can the README.md usage be updated with more examples on the usage of the parameters. Note -DestinationLibrary must be used with RemoveSharingFileAccess.

ASSISTANT: Sure, I can update the README.md file to include more examples and a note about the usage of the -DestinationLibrary and -RemoveSharingFileAccess parameters. Here is the proposed change:

<source>markdown
README.md
<<<<<<< HEAD
## Usage
```powershell
.\get-sp-rp.ps1 -SiteUrl "https://yoursharepointsite.com" -ListName "Your List Name" -DestinationLibrary "Your Destination Library"
```
Replace `"https://yoursharepointsite.com"` with the URL of your SharePoint site, `"Your List Name"` with the name of the list you want to get the shared links from, and `"Your Destination Library"` with the name of your destination library.

Please note that you need to have the SharePoint PnP PowerShell module installed and you need to have access to the SharePoint site.
=======
## Usage
```powershell
.\get-sp-rp.ps1 -SiteUrl "https://yoursharepointsite.com" -ListName "Your List Name" -DestinationLibrary "Your Destination Library"
.\get-sp-rp.ps1 -SiteUrl "https://yoursharepointsite.com" -ListName "Your List Name" -DestinationLibrary "Your Destination Library" -RemoveSharingFileAccess $true
.\get-sp-rp.ps1 -SiteUrl "https://yoursharepointsite.com" -ListName "Your List Name" -DestinationLibrary "Your Destination Library" -ReportOutput "myreport.csv"
.\get-sp-rp.ps1 -SiteUrl "https://yoursharepointsite.com" -ListName "Your List Name" -DestinationLibrary "Your Destination Library" -LogFile "mylog.txt"
.\get-sp-rp.ps1 -SiteUrl "https://yoursharepointsite.com" -ListName "Your List Name" -DestinationLibrary "Your Destination Library" -UsePreviousOutput $true
.\get-sp-rp.ps1 -SiteUrl "https://yoursharepointsite.com" -ListName "Your List Name" -DestinationLibrary "Your Destination Library" -InputFile "myinputfile.txt"
```
Replace `"https://yoursharepointsite.com"` with the URL of your SharePoint site, `"Your List Name"` with the name of the list you want to get the shared links from, and `"Your Destination Library"` with the name of your destination library. Replace `"myreport.csv"`, `"mylog.txt"`, and `"myinputfile.txt"` with your desired output report file, log file, and input file respectively.

Please note that you need to have the SharePoint PnP PowerShell module installed and you need to have access to the SharePoint site.

**Note:** The `-DestinationLibrary` parameter must be used with the `-RemoveSharingFileAccess` parameter.
>>>>>>> updated
</source>