aider: Removed redundant line about needing the SharePoint PnP PowerShell module and access to the SharePoint site in the README.md file.
# Aider chat conversation:

USER: The last line i.e Please note that you need to have the SharePoint PnP PowerShell module installed and you need to have access to the SharePoint site. Can be removed as this is already mentioned in the Usage.

ASSISTANT: To remove the redundant line from the README.md file, we need to modify the last paragraph. Here is the *edit block* for the changes:

<source>markdown
README.md
<<<<<<< HEAD
Replace `"https://yoursharepointsite.com"` with the URL of your SharePoint site, `"Your List Name"` with the name of the list you want to get the shared links from, and `"Your Destination Library"` with the name of your destination library. Replace `"myreport.csv"`, `"mylog.txt"`, and `"myinputfile.txt"` with your desired output report file, log file, and input file respectively.

Please note that you need to have the SharePoint PnP PowerShell module installed and you need to have access to the SharePoint site.

**Note:** The `-DestinationLibrary` parameter must be used with the `-RemoveSharingFileAccess` parameter.
=======
Replace `"https://yoursharepointsite.com"` with the URL of your SharePoint site, `"Your List Name"` with the name of the list you want to get the shared links from, and `"Your Destination Library"` with the name of your destination library. Replace `"myreport.csv"`, `"mylog.txt"`, and `"myinputfile.txt"` with your desired output report file, log file, and input file respectively.

**Note:** The `-DestinationLibrary` parameter must be used with the `-RemoveSharingFileAccess` parameter.
>>>>>>> updated
</source>