aider: Added information about the importance of the input file for speeding up the processing of the script.
# Aider chat conversation:

USER: Can you provide the importance of the inputfile under the Usage. The input file allow you to speed up the processing of the script the 2nd time you run as only we process the list of files with expiration date instead of running against all the documents in the document library. That is huge time saving factor if the document library is many documents.

ASSISTANT: Sure, I can add that information to the `README.md` file. Here's the proposed change:

<source>markdown
README.md
<<<<<<< HEAD
.\get-sp-rp.ps1 -SiteUrl "https://yoursharepointsite.com" -ListName "Your List Name" -DestinationLibrary "Your Destination Library" -RemoveSharingFileAccess $true -InputFile "myinputfile.txt" -ReportOutput "myreport.csv"
```
Replace `"https://yoursharepointsite.com"` with the URL of your SharePoint site, `"Your List Name"` with the name of the list you want to get the shared links from, and `"Your Destination Library"` with the name of your destination library. Replace `"myreport.csv"` and `"myinputfile.txt"` with your desired output report file and input file respectively.
=======
.\get-sp-rp.ps1 -SiteUrl "https://yoursharepointsite.com" -ListName "Your List Name" -DestinationLibrary "Your Destination Library" -RemoveSharingFileAccess $true -InputFile "myinputfile.txt" -ReportOutput "myreport.csv"
```
Replace `"https://yoursharepointsite.com"` with the URL of your SharePoint site, `"Your List Name"` with the name of the list you want to get the shared links from, and `"Your Destination Library"` with the name of your destination library. Replace `"myreport.csv"` and `"myinputfile.txt"` with your desired output report file and input file respectively.

The `-InputFile` parameter allows you to speed up the processing of the script the second time you run it, as it will only process the list of files with an expiration date instead of running against all the documents in the document library. This can be a significant time-saving factor if the document library contains many documents.
>>>>>>> updated
</source>