aider: Modified script to use `Get-PnPFile` instead of `Get-PnPListItem` to retrieve items by URL.
# Aider chat conversation:

USER: Get review your code, Get-PnpListItem retrieves items by ID, however ID is a int32 type and maynot access the RelativeURL as this is contains string.

ASSISTANT: You're correct. The `Get-PnPListItem` cmdlet retrieves items by ID, which is an integer. The `RelativeURL` is a string and cannot be used as an ID.

To fetch items by URL, we can use the `Get-PnPFile` cmdlet instead. This cmdlet retrieves a file object given a server-relative URL.

Here's how we can modify the script:

<source>powershell
get-sp-rp.ps1
<<<<<<< HEAD
    if ($InputFile) {
        # If an input file is provided, read the file and process only those items
        $ListItems = Import-Csv $InputFile | ForEach-Object {
            Get-PnPListItem -List $ListName -Id $_.RelativeURL -Fields "FileLeafRef", "FileRef", "File_x0020_Type"
        }
    } else {
        # If no input file is provided, fetch all items
        $ListItems = Get-PnPListItem -List $ListName -PageSize 2000 -Fields "FileLeafRef", "FileRef", "File_x0020_Type"
    }
=======
    if ($InputFile) {
        # If an input file is provided, read the file and process only those items
        $ListItems = Import-Csv $InputFile | ForEach-Object {
            Get-PnPFile -Url $_.RelativeURL -AsListItem -Fields "FileLeafRef", "FileRef", "File_x0020_Type"
        }
    } else {
        # If no input file is provided, fetch all items
        $ListItems = Get-PnPListItem -List $ListName -PageSize 2000 -Fields "FileLeafRef", "FileRef", "File_x0020_Type"
    }
>>>>>>> updated
</source>

This change will make the script read the CSV file and use the `RelativeURL` column as the URL for `Get-PnPFile`. The `-AsListItem` parameter is used to return the file as a list item, which allows us to access the same fields as before.