aider: Changed parameters for -ListName to -SL and -DestinationLibrary to -DL in get-sp-rp.ps1 and set-sp-rp.ps1 scripts.
# Aider chat conversation:

USER: Can you change the parameters for -ListName to -SL i.e Source Library and -DestinationLibrary to -DL for the scripts

ASSISTANT: Sure, I can help with that. We will need to modify the `get-sp-rp.ps1` and `set-sp-rp.ps1` scripts.

Here are the changes:

For `get-sp-rp.ps1`:

<source>powershell
get-sp-rp.ps1
<<<<<<< HEAD
    [Parameter(Mandatory=$true)]
    [string]$ListName,
=======
    [Parameter(Mandatory=$true)]
    [string]$SL,
>>>>>>> updated
</source>

For `set-sp-rp.ps1`:

<source>powershell
set-sp-rp.ps1
<<<<<<< HEAD
    [Parameter(Mandatory=$true)]
    [string]$ListName,
    [Parameter(Mandatory=$true)]
    [string]$DestinationLibrary,
=======
    [Parameter(Mandatory=$true)]
    [string]$SL,
    [Parameter(Mandatory=$true)]
    [string]$DL,
>>>>>>> updated
</source>

Please note that you will also need to replace all instances of `$ListName` with `$SL` and `$DestinationLibrary` with `$DL` in the scripts.