aider: Refactored Get-PnPFile command to use Get-PnPListItem to retrieve specific fields for each file.
# Aider chat conversation:

USER: in the Get-PnPFile -Url $_.RelativeURL -AsListItem -Fields "FileLeafRef", "FileRef", "File_x0020_Type" there isn't a parameter for -Fields

ASSISTANT: It seems like you're trying to use the `-Fields` parameter with the `Get-PnPFile` cmdlet in PowerShell, but this cmdlet doesn't support the `-Fields` parameter.

If you want to retrieve specific fields from a file, you might need to use a different cmdlet or approach. For example, you could use `Get-PnPListItem` to retrieve list items and their fields.

However, without the actual script file, I can't provide a precise solution. Could you please add the script file to the chat so I can suggest the appropriate changes?

USER: the script is in the get-sp-rp.ps1

ASSISTANT: I see, thank you for pointing that out.

To fix the issue, we need to modify the `get-sp-rp.ps1` file.

The `Get-PnPFile` cmdlet doesn't support the `-Fields` parameter. Instead, we can use the `Get-PnPListItem` cmdlet to retrieve the list items and their fields.

Here are the changes:

```powershell
get-sp-rp.ps1
<<<<<<< HEAD
    if ($InputFile) {
        # If an input file is provided, read the file and process only those items
        $ListItems = Import-Csv $InputFile | ForEach-Object {
            Get-PnPFile -Url $_.RelativeURL -AsListItem -Fields "FileLeafRef", "FileRef", "File_x0020_Type"
        }
    } else {
        # If no input file is provided, fetch all items
        $ListItems = Get-PnPListItem -List $ListName -PageSize 2000 -Fields "FileLeafRef", "FileRef", "File_x0020_Type"
    }
=======
    if ($InputFile) {
        # If an input file is provided, read the file and process only those items
        $ListItems = Import-Csv $InputFile | ForEach-Object {
            $file = Get-PnPFile -Url $_.RelativeURL -AsListItem
            Get-PnPListItem -List $file.ListItemAllFields.ParentList -Id $file.ListItemAllFields.Id -Fields "FileLeafRef", "FileRef", "File_x0020_Type"
        }
    } else {
        # If no input file is provided, fetch all items
        $ListItems = Get-PnPListItem -List $ListName -PageSize 2000 -Fields "FileLeafRef", "FileRef", "File_x0020_Type"
    }
>>>>>>> updated
```

In the updated code, we first get the file as a list item using `Get-PnPFile -Url $_.RelativeURL -AsListItem`. Then, we use `Get-PnPListItem` to retrieve the list item and its fields. We use the `ListItemAllFields.ParentList` property of the file to get the parent list of the file, and the `ListItemAllFields.Id` property to get the ID of the list item.